reorganize queries
