+++ src/pages/CreateEvent/queries/ActivitiesQuery.js
@@ -1,0 +1,35 @@
+import gql from 'graphql-tag';
+
+export const CREATE_ACTIVITY = gql`
+  mutation CreateActivity(
+    $name: String!
+    $startDate: String!
+    $startTime: String!
+    $location: String!
+    $type: String!
+    $details: String
+    $event_id: ID
+  ) {
+    createActivity(
+      data: {
+        name: $name
+        startDate: $startDate
+        startTime: $startTime
+        location: $location
+        type: $type
+        details: $details
+        event: { connect: { id: $event_id } }
+      }
+    ) {
+      id
+      name
+      startDate
+      startTime
+      location
+      type
+      event {
+        title
+      }
+    }
+  }
+`;
