+++ src/pages/CreateEvent/queries/EventsQuery.js
@@ -50,38 +50,3 @@ export const UPDATE_EVENT = gql`
     }
   }
 `;
-
-// Creating an activity
-export const CREATE_ACTIVITY = gql`
-  mutation CreateActivity(
-    $name: String!
-    $startDate: String!
-    $startTime: String!
-    $location: String!
-    $type: String!
-    $details: String
-    $event_id: ID
-  ) {
-    createActivity(
-      data: {
-        name: $name
-        startDate: $startDate
-        startTime: $startTime
-        location: $location
-        type: $type
-        details: $details
-        event: { connect: { id: $event_id } }
-      }
-    ) {
-      id
-      name
-      startDate
-      startTime
-      location
-      type
-      event {
-        title
-      }
-    }
-  }
-`;
